backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.0
httpx==0.28.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.2.1
pypdf==4.3.1
beautifulsoup4==4.12.3
